+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * ROWNUM() and ROW_NUMBER() functions
 * Promised version of alasql() - alasql.async() (based on es6-promises)
 * SELECT * FROM Json
+* SEARCH COMMA selector
 
 ### 0.1.5 "San Marino" (12.05.2015 - 12.05.2015)
 * Added Meteor package (agershun:alasql) - still does not work - skeleton